requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.2
python-docx==1.1.2
requests==2.32.3
typing_extensions==4.12.1
urllib3==2.2.1
